+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -17,16 +17,19 @@ import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Random;
+import java.util.RandomAccess;
 import java.util.Set;
+import java.util.TreeSet;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 import org.checkerframework.checker.index.qual.NonNegative;
 import org.checkerframework.checker.index.qual.Positive;
 import org.checkerframework.checker.lock.qual.GuardSatisfied;
 import org.checkerframework.checker.nullness.qual.KeyFor;
+import org.checkerframework.checker.nullness.qual.KeyForBottom;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import org.checkerframework.checker.nullness.qual.UnknownKeyFor;
 import org.checkerframework.dataflow.qual.Pure;
 
 /** Utility functions for Collections, ArrayList, Iterator, and Map. */
@@ -59,19 +62,111 @@ public final class CollectionsPlume {
     return result;
   }
 
-  // This should perhaps be named withoutDuplicates to emphasize that
-  // it does not side-effect its argument.
   /**
    * Return a copy of the list with duplicates removed. Retains the original order.
    *
    * @param <T> type of elements of the list
    * @param l a list to remove duplicates from
    * @return a copy of the list with duplicates removed
+   * @deprecated use {@link withoutDuplicates}
    */
+  @Deprecated // 2021-03-28
   public static <T> List<T> removeDuplicates(List<T> l) {
     HashSet<T> hs = new LinkedHashSet<>(l);
     List<T> result = new ArrayList<>(hs);
     return result;
+  }
+
+  /**
+   * Returns a list with the same contents as its argument, but without duplicates. May return its
+   * argument if its argument has no duplicates, but is not guaranteed to do so.
+   *
+   * @param <T> the type of elements in {@code values}
+   * @param values a list of values
+   * @return the values, with duplicates removed
+   */
+  public static <T extends Comparable<T>> List<T> withoutDuplicates(List<T> values) {
+    // This adds O(n) time cost, and has the benefit of sometimes avoiding allocating a TreeSet.
+    if (isSortedNoDuplicates(values)) {
+      return values;
+    }
+
+    Set<T> set = new TreeSet<>(values);
+    if (values.size() == set.size()) {
+      return values;
+    } else {
+      return new ArrayList<>(set);
+    }
+  }
+
+  /**
+   * Returns true if the given list is sorted.
+   *
+   * @param <T> the component type of the list
+   * @param values a list
+   * @return true if the list is sorted
+   */
+  public static <T extends Comparable<T>> boolean isSorted(List<T> values) {
+    if (values.isEmpty() || values.size() == 1) {
+      return true;
+    }
+
+    if (values instanceof RandomAccess) {
+      // Per the Javadoc of RandomAccess, an indexed for loop is faster than a foreach loop.
+      int size = values.size();
+      for (int i = 0; i < size - 1; i++) {
+        if (values.get(i).compareTo(values.get(i + 1)) > 0) {
+          return false;
+        }
+      }
+      return true;
+    } else {
+      Iterator<T> iter = values.iterator();
+      T previous = iter.next();
+      while (iter.hasNext()) {
+        T current = iter.next();
+        if (previous.compareTo(current) > 0) {
+          return false;
+        }
+        previous = current;
+      }
+      return true;
+    }
+  }
+
+  /**
+   * Returns true if the given list is sorted and has no duplicates
+   *
+   * @param <T> the component type of the list
+   * @param values a list
+   * @return true if the list is sorted and has no duplicates
+   */
+  public static <T extends Comparable<T>> boolean isSortedNoDuplicates(List<T> values) {
+    if (values.isEmpty() || values.size() == 1) {
+      return true;
+    }
+
+    if (values instanceof RandomAccess) {
+      // Per the Javadoc of RandomAccess, an indexed for loop is faster than a foreach loop.
+      int size = values.size();
+      for (int i = 0; i < size - 1; i++) {
+        if (values.get(i).compareTo(values.get(i + 1)) >= 0) {
+          return false;
+        }
+      }
+      return true;
+    } else {
+      Iterator<T> iter = values.iterator();
+      T previous = iter.next();
+      while (iter.hasNext()) {
+        T current = iter.next();
+        if (previous.compareTo(current) >= 0) {
+          return false;
+        }
+        previous = current;
+      }
+      return true;
+    }
   }
 
   /** All calls to deepEquals that are currently underway. */
@@ -168,33 +263,78 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Applies the function to each element of the given collection, producing a list of the results.
+   * Applies the function to each element of the given iterable, producing a list of the results.
    *
    * <p>The point of this method is to make mapping operations more concise. Import it with
    *
    * <pre>import static org.plumelib.util.CollectionsPlume.mapList;</pre>
    *
-   * This method is just like {@link #transform}, but with the arguments in the other order.
-   *
-   * @param <FROM> the type of elements of the given collection
+   * @param <FROM> the type of elements of the given iterable
    * @param <TO> the type of elements of the result list
    * @param f a function
-   * @param c a collection
-   * @return a list of the results of applying {@code f} to the elements of {@code list}
+   * @param iterable an iterable
+   * @return a list of the results of applying {@code f} to the elements of {@code iterable}
    */
-  public static <FROM, TO> List<TO> mapList(Function<? super FROM, TO> f, Collection<FROM> c) {
-    // This implementation uses a for loop and is likely more efficient than using streams, both
-    // because it doesn't create stream objects and because it creates an ArrayList of the
-    // appropriate size.
-    List<TO> result = new ArrayList<>(c.size());
-    for (FROM elt : c) {
+  public static <
+          @KeyForBottom FROM extends @UnknownKeyFor Object,
+          @KeyForBottom TO extends @UnknownKeyFor Object>
+      List<TO> mapList(Function<? super FROM, ? extends TO> f, Iterable<FROM> iterable) {
+    List<TO> result;
+
+    if (iterable instanceof RandomAccess) {
+      // Per the Javadoc of RandomAccess, an indexed for loop is faster than a foreach loop.
+      List<FROM> list = (List<FROM>) iterable;
+      int size = list.size();
+      result = new ArrayList<>(size);
+      for (int i = 0; i < size; i++) {
+        result.add(f.apply(list.get(i)));
+      }
+      return result;
+    }
+
+    if (iterable instanceof Collection) {
+      result = new ArrayList<>(((Collection<?>) iterable).size());
+    } else {
+      result = new ArrayList<>(); // no information about size is available
+    }
+    for (FROM elt : iterable) {
       result.add(f.apply(elt));
     }
     return result;
   }
 
   /**
-   * Applies the function to each element of the given collection, producing a list of the results.
+   * Applies the function to each element of the given array, producing a list of the results.
+   *
+   * <p>This produces a list rather than an array because it is problematic to create an array with
+   * generic compontent type.
+   *
+   * <p>The point of this method is to make mapping operations more concise. Import it with
+   *
+   * <pre>import static org.plumelib.util.CollectionsPlume.mapList;</pre>
+   *
+   * @param <FROM> the type of elements of the given array
+   * @param <TO> the type of elements of the result list
+   * @param f a function
+   * @param a an array
+   * @return a list of the results of applying {@code f} to the elements of {@code a}
+   */
+  public static <
+          @KeyForBottom FROM extends @UnknownKeyFor Object,
+          @KeyForBottom TO extends @UnknownKeyFor Object>
+      List<TO> mapList(Function<? super FROM, ? extends TO> f, FROM[] a) {
+    int size = a.length;
+    List<TO> result = new ArrayList<>(size);
+    for (int i = 0; i < size; i++) {
+      result.add(f.apply(a[i]));
+    }
+    return result;
+  }
+
+  /**
+   * Applies the function to each element of the given iterable, producing a list of the results.
+   * This is just like {@link #mapList(Function, Iterable)}, but with the arguments in the opposite
+   * order.
    *
    * <p>The point of this method is to make mapping operations more concise. Import it with
    *
@@ -204,12 +344,15 @@ public final class CollectionsPlume {
    *
    * @param <FROM> the type of elements of the given collection
    * @param <TO> the type of elements of the result list
-   * @param c a collection
+   * @param iterable an iterable
    * @param f a function
    * @return a list of the results of applying {@code f} to the elements of {@code list}
    */
-  public static <FROM, TO> List<TO> transform(Collection<FROM> c, Function<? super FROM, TO> f) {
-    return c.stream().map(f).collect(Collectors.toList());
+  public static <
+          @KeyForBottom FROM extends @UnknownKeyFor Object,
+          @KeyForBottom TO extends @UnknownKeyFor Object>
+      List<TO> transform(Iterable<FROM> iterable, Function<? super FROM, ? extends TO> f) {
+    return mapList(f, iterable);
   }
 
   ///////////////////////////////////////////////////////////////////////////
@@ -260,6 +403,22 @@ public final class CollectionsPlume {
     List<T> result = new ArrayList<>(list.size() + 1);
     result.addAll(list);
     result.add(lastElt);
+    return result;
+  }
+
+  /**
+   * Concatenates two lists into a new list.
+   *
+   * @param <T> the type of the list elements
+   * @param list1 the first list
+   * @param list2 the second list
+   * @return a new list containing the contents of the given lists, in order
+   */
+  @SuppressWarnings("unchecked")
+  public static <T> List<T> concatenate(Collection<T> list1, Collection<T> list2) {
+    List<T> result = new ArrayList<>(list1.size() + list2.size());
+    result.addAll(list1);
+    result.addAll(list2);
     return result;
   }